% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenerate.R
\name{autogenerate}
\alias{autogenerate}
\title{Autogenerate a datastorr interface}
\usage{
autogenerate(repo, read, filename = NULL, name = basename(repo),
  roxygen = TRUE)
}
\arguments{
\item{repo}{Name of the repo on github (in username/repo format)}

\item{read}{\emph{name} of a function to read the data.  Do not
give the function itself!}

\item{filename}{Name of the file to read.  If not given, then the
single file in a release will be read (but you will need to
provide a filename on upload).  If given, you cannot change the
filename ever as all releases will be assumed to have the same
filename.}

\item{name}{Name of the dataset, used in generating the functions.
If omitted the repo name is used.}

\item{roxygen}{Include roxygen headers for the functions?}
}
\description{
Autogenerate an datastorr interface for a package.  The idea is to
run this function and save the resulting code in a file in your
package.  Then users will be able to download data and you will be
able to relase data easily.
}
\details{
In addition to running this, you will need to add \code{datastorr}
to the \code{Imports:} section of your DESCRIPTION.  To upload
files you will need to set your \code{GITHUB_TOKEN} environment
variable.  These steps will be described more fully in a vignette.
}

