% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_auth.R
\name{datastorr_auth}
\alias{datastorr_auth}
\alias{setup_github_token}
\title{datastorr/GitHub authentication}
\usage{
datastorr_auth(required = FALSE, key = NULL, secret = NULL,
  cache = TRUE)

setup_github_token(path = "~/.Renviron")
}
\arguments{
\item{required}{Is authentication required?  Reading from public
repositories does not require authentication so there's no point
worrying if we can't get it.  datastorr will set this when
appropriate internally.}

\item{key, secret}{The application key and secret.  If \code{NULL},
uses datastorr's key.  But if you have your own application feel
free to replace these with your own.}

\item{cache}{Logical, indicating whether we should cache the
token.  If \code{TRUE}, the token will be cached at
\code{\link{datastorr_auth}()}, so that it is accessible to all
datastorr usages.  Note that this is affected by the
\code{datastorr.path} global option.  Alternatively, set
\code{FALSE} to do no caching and be prompted each session or a
string to choose your own filename.  Or set the
\code{GITHUB_TOKEN} or \code{GITHUB_PAT} environment variables
to use a token rather than OAuth.}

\item{path}{Path to environment file; the default is the user
environment variable file which is usually a good choice.}
}
\description{
Authentication for accessing GitHub.  This will first look for a
GitHub personal token (stored in the \code{GITHUB_TOKEN} or
\code{GITHUB_PAT} environment variables, and then try
authenticating with OAuth.
}
\details{
Run this \code{datastorr_auth} function to force setting up
authentication with OAuth.  Alternatively, run
\code{setup_github_token} to set up a personal access token.
Either can be revoked at any time
\url{https://github.com/settings/tokens} to revke a personal
access token and \url{https://github.com/settings/applications} to
revoke the OAuth token.
}

