% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{datastorr_path}
\alias{datastorr_path}
\title{Location of datastorr files}
\usage{
datastorr_path(repo = NULL)
}
\arguments{
\item{repo}{An optional repo (of the form \code{user/repo}, though
this is not checked).}
}
\description{
Location of datastorr files.  This is determined by
\code{rappdirs} using the \code{user_data_dir} function.
Alternatively, if the option \code{datastorr.path} is set, that is
used for the base path.  The path to data from an actual repo is
stored in a subdirectory under this directory.
}
\details{
Files in this directory can be deleted at will (e.g., running
\code{unlink(datastorr_path(), recursive = TRUE)} will delete all
files that datstorr has ever downloaded.  The only issue here is
that the OAuth token (used to authenticate with GitHub) is also
stored in this directory.
}
