% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_release_del}
\alias{github_release_del}
\title{Delete version}
\usage{
github_release_del(info, version)
}
\arguments{
\item{info}{Result of running \code{github_release_info}}

\item{version}{Version to delete.  If \code{NULL} it will delete the entire storr}
}
\description{
Delete a local copy of a version (or the entire thing).
}

