% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_release_get}
\alias{github_release_get}
\title{Get data}
\usage{
github_release_get(info, version = NULL)
}
\arguments{
\item{info}{Result of running \code{github_release_info}}

\item{version}{Version to fetch.  If \code{NULL} it will get the
current version as returned by
\code{github_release_version_current()}}
}
\description{
Get a version of a data set, downloading it if necessary.
}
