% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_release_info}
\alias{github_release_info}
\title{Github release information}
\usage{
github_release_info(repo, read, private = FALSE, filename = NULL,
  path = NULL)
}
\arguments{
\item{repo}{Name of the repo in \code{username/repo} format.}

\item{read}{Function to read the file.  See Details.}

\item{private}{Is the repository private?  If so authentication
will be required for all actions.  Setting this is optional but
will result in better error messages because of the way GitHub
returns not found/404 (rather than forbidden/403) errors when
accessing private repositories without authorisation.}

\item{filename}{Optional filename.  If omitted, all files in the
release can be used.  If the filename contains a star ("*") it
will be treated as a filename glob.  So you can do
\code{filename="*.csv"} to match all csv files (dynamically
computed on each release).}

\item{path}{Optional path in which to store the data.  If omitted
we use \code{\link{datastorr_path}} to generate a reasonable
path.}
}
\description{
Information to describe how to process github releases
}
\details{
The simplest case is where the data are stored in a single file
attached to the release (this is different to the zip/tar.gz files
that the web interface displays).  For example, a single csv file.
In that case the filename argument can be safely ommited and we'll
work it out based on the filename.
}

