% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_release_versions}
\alias{github_release_version_current}
\alias{github_release_versions}
\title{Get release versions}
\usage{
github_release_versions(info, local = TRUE)

github_release_version_current(info, local = TRUE)
}
\arguments{
\item{info}{Result of running \code{github_release_info}}

\item{local}{Should we return local (TRUE) or github (FALSE)
version numbers?  Github version numbers are pulled once per
session only.  The exception is for
\code{github_release_version_current} which when given
\code{local=TRUE} will fall back on trying github if there are
no local versions.}
}
\description{
Get release versions
}
\author{
Rich FitzJohn
}

