% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{release}
\alias{release}
\title{Release data to a datastorr repository}
\usage{
release(repo, version, description = NULL, filename = NULL, path = NULL,
  metadata = "datastorr.json", branch = "master", private = FALSE,
  refetch = FALSE, target = NULL, ignore_dirty = FALSE,
  yes = !interactive())
}
\arguments{
\item{repo}{Either a github repo in the form
\code{<username>/<repo>} (e.g.,
\code{"richfitz/data"} or the path to a json file
on your filesystem.}

\item{version}{A version number for the new version.  Should be of
the form x.y.z, and may or may not contain a leading "v" (one
will be added in any case).}

\item{description}{Optional text description for the release.  If
this is omitted then GitHub will display the commit message from
the commit that the release points at.}

\item{filename}{Filename to upload; optional if in
\code{datastorr.json}.  If listed, \code{filename} can be
different but the file will be renamed on uploading.  If given
but not in \code{info}, the uploaded file will be
\code{basename(filename)} (i.e., the directory will be
stripped).}

\item{path}{The path to store the data at.  Using \code{NULL} will}

\item{metadata}{The name of the metadata file within the repo (if
\code{repo} refers to a github repo.  The default is
\code{datastorr.json} at the root of the repository, but any
other filename can be used.}

\item{branch}{The branch in the repo to use.  Default is
\code{master}.}

\item{private}{A logical indicating if the repository is private
and therefor if authentication will be needed to access it.}

\item{refetch}{Refetch the metadata file even if it has already
been downloaded previously.}

\item{target}{The SHA or tag to attach the release to.  By
default, will use the current HEAD, which is typically what you
want to do.}

\item{ignore_dirty}{Ignore non-checked in files?  By default, your
repository is expected to be in a clean state, though files not
known to git are ignored (as are files that are ignored by git).
But you must have no uncommited changes or staged but uncommited
files.}

\item{yes}{Skip the confirmation prompt?  Only prompts if
interactive.}
}
\description{
Create a relase for a simple datastorr (i.e., non-package based).
}
