% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_download.R
\name{fetch_hook_download}
\alias{fetch_hook_download}
\title{Hook for downloading files}
\usage{
fetch_hook_download(furl, fread)
}
\arguments{
\item{furl}{Function to convert \code{key, namespace} into a URL.}

\item{fread}{Function for converting \code{filename} into an R
object.}
}
\description{
Storr hook for downloading files as an external resource
}
\seealso{
\code{\link{driver_external}}
}

